Refactor App, Calculator, index and delete routes
