+++ package.json
@@ -4,7 +4,6 @@
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "big.js": "^6.1.1",
     "enzyme-adapter-react-16": "^1.15.6",
@@ -44,6 +43,7 @@
     "@babel/eslint-parser": "^7.14.7",
     "@babel/plugin-syntax-jsx": "^7.14.5",
     "@babel/preset-react": "^7.14.5",
+    "@testing-library/react": "^12.0.0",
     "babel-preset-stage-0": "^6.24.1",
     "enzyme": "^3.11.0",
     "eslint": "^7.31.0",
@@ -63,5 +63,4 @@
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "^3.20.1"
   }
-  
 }
